requirements.txt
torch>=2.3.0
torchvision>=0.18.0
monai>=1.3.1
nnunetv2>=2.5.1
optuna>=3.6.1
cleanlab>=2.6.5
mrcfile>=1.5.1
numpy>=1.26.4
scikit-learn>=1.4.2
scipy>=1.13.1
pandas>=2.2.2
pyyaml>=6.0.1
tqdm>=4.66.4
matplotlib>=3.8.4
